% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_plots.R
\name{catplot}
\alias{catplot}
\title{Category plots}
\usage{
catplot(categories, cat_labels, ...)
}
\arguments{
\item{categories}{Category orders produced from \code{rasch}}

\item{cat_labels}{Optional category labels in the style of c(`colname1` = "newlabel1", `colname2` = "newlabel2").
If not specified, labels will be taken from the category column names}

\item{\dots}{Other arguments to pass onto \code{\link{ggplot}}}
}
\description{
Category plots
}
\author{
Tara Valladares <tls8vx at virginia.edu>
}
