% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_fit.R
\name{rasch_fit}
\alias{rasch_fit}
\title{Calculate Rasch fit statistics}
\usage{
rasch_fit(model, groups = FALSE)
}
\arguments{
\item{model}{The object outputted by the \code{\link{emoxicon}} \code{\link{rasch}} function.}

\item{groups}{Logical. Should summary fit statistics be calculated across
the individual group models, if present? The default is FALSE. Running fit statistics
on many models can be slow.}
}
\description{
This function serves as a wrapper to
calculate Rasch person and item fit statistics based
on the \code{\link{eRm}} fit functions.
}
\examples{
# Load the tinytrolls data
data(tinytrolls)

# Use the emoxicon function
\dontrun{
emotions.tinytrolls <- emoxicon(text = tinyTrolls$content, lexicon = emotions)
# Recode the variables to 0 (below mean) or 1 (equal to or above mean)
emotions.rasch <- as.data.frame(apply(emotions.tinytrolls[,-c(1:2)],2,function(x) ifelse(x<mean(x),0,1)))
# Apply the Rasch Model
rm.tinytrolls <- rasch(scores = emotions.tinytrolls)
# Use the rasch_fit function
fit.tinytrolls <- rasch_fit(rm.tinytrolls)
}


}
\author{
Tara Valladares <tls8vx at virginia.edu>
}
