% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryCatchWE.R
\name{tryCatch.W.E}
\alias{tryCatch.W.E}
\title{Catch errors and warnings}
\usage{
tryCatch.W.E(expr)
}
\arguments{
\item{expr}{R expression to evaluate.}
}
\value{
The function returns all warning and error messages and the function output.
}
\description{
Catch any errors and warnings and then rerun the function if necessary.
The function returns all messages and the function output.
}
\examples{
# Load the tinytrolls data
data(tinytrolls)

# Use the emoxicon function
\dontrun{
warns.errs.tinytrolls <- tryCatch.W.E(emoxicon(text = tinyTrolls$content, lexicon = emotions))
}
}
\author{
Tara Valladares <tls8vx at virginia.edu>
}
