% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch.R
\name{rasch}
\alias{rasch}
\title{Estimation of Rasch Models for Emoxicon}
\usage{
rasch(scores, groups = NULL, return_models = TRUE, ...)
}
\arguments{
\item{scores}{A dataframe containing the emotion scores}

\item{groups}{An optional vector of group identifers the same length as \code{scores}.
If provided, individual Rasch models will be run by group.}

\item{return_models}{logical. Should all models for the groups be returned?}

\item{...}{Additional arguments to be passed to \code{\link[eRm]{RM}}}
}
\description{
Runs a Rasch model using the emotion scores from \code{emoxicon}.
If the data is not already dichotomous, a mean split will be performed.
}
\details{
When the data is generated using \code{emoxicon} and the default emotions lexicon,
the entire emoxicon object can be entered.
}
\examples{

\donttest{
#examples here
}
}
\references{

}
\seealso{
\code{\link{emoxicon}}, where the emotion scores are generated.
}
\author{
Tara Valladares <tls8vx at virginia.edu>, Hudson F. Golino <hfg9s at virginia.edu>
}
