% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_fit.R
\name{rasch_fit}
\alias{rasch_fit}
\title{Return Rasch fit statistics}
\usage{
rasch_fit(model, groups = FALSE)
}
\arguments{
\item{model}{The object outputted by the \code{emoxicon} \code{rasch} function.}

\item{groups}{Logical. Should summary fit statistics be calculated across
the individual group models, if present? The default is FALSE. Running fit statistics
on many models can be slow.}
}
\description{
This function calculates Rasch model person and item fit statistics.
}
\author{
Tara Valladares <tls8vx at virginia.edu>
}
